refactored docblock completor to be tolerant
