+++ lib/Completion/Tests/Integration/Bridge/TolerantParser/WorseReflection/DocblockCompletorTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Phpactor\Completion\Tests\Integration\Bridge\WorseReflection\Completor;
+namespace Phpactor\Completion\Tests\Integration\Bridge\TolerantParser\WorseReflection;
 
 use Generator;
+use Microsoft\PhpParser\Parser;
 use PHPUnit\Framework\TestCase;
-use Phpactor\Completion\Bridge\WorseReflection\Completor\DocblockCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\Completor\DocblockCompletor;
 use Phpactor\Completion\Core\Suggestion;
 use Phpactor\Name\FullyQualifiedName;
 use Phpactor\ReferenceFinder\Search\NameSearchResult;
@@ -29,7 +30,9 @@ class DocblockCompletorTest extends TestCase
         ];
 
         [$source, $offset] = ExtractOffset::fromSource($source);
+        $node = (new Parser())->parseSourceFile($source)->getDescendantNodeAtPosition((int)$offset);
         $suggestions = iterator_to_array((new DocblockCompletor(new PredefinedNameSearcher($results)))->complete(
+            $node,
             TextDocumentBuilder::create($source)->build(),
             ByteOffset::fromInt((int)$offset)
         ));
@@ -68,7 +71,7 @@ class DocblockCompletorTest extends TestCase
         ];
 
         yield 'bare ampersand' => [
-            '   *    @<> */',
+            '   *    @<>',
             DocblockCompletor::SUPPORTED_TAGS,
         ];
